+++ pool.go
@@ -6,11 +6,15 @@ import (
 	"sync/atomic"
 
 	"golang.org/x/net/context"
-	"gopkg.in/fatih/pool.v2"
 )
 
 var (
 	errPoolClosed = errors.New("rethinkdb: pool is closed")
+)
+
+const (
+	poolIsNotClosed int32 = 0
+	poolIsClosed    int32 = 1
 )
 
 // A Pool is used to store a pool of connections to a single RethinkDB server
@@ -20,9 +24,9 @@ type Pool struct {
 
 	conns   []*Connection
 	pointer int32
+	closed  int32
 
-	mu     sync.RWMutex // protects following fields
-	closed bool
+	mu sync.Mutex // protects lazy creating connections
 }
 
 // NewPool creates a new connection pool for the given host
@@ -41,7 +45,7 @@ func NewPool(host Host, opts *ConnectOpts) (*Pool, error) {
 
 	conns := make([]*Connection, maxOpen)
 	var err error
-	for i := range conns {
+	for i := 0; i < opts.InitialCap; i++ {
 		conns[i], err = NewConnection(host.String(), opts)
 		if err != nil {
 			return nil, err
@@ -53,13 +57,14 @@ func NewPool(host Host, opts *ConnectOpts) (*Pool, error) {
 		pointer: -1,
 		host:    host,
 		opts:    opts,
+		closed:  poolIsNotClosed,
 	}, nil
 }
 
 // Ping verifies a connection to the database is still alive,
 // establishing a connection if necessary.
 func (p *Pool) Ping() error {
-	_, _, err := p.conn()
+	_, err := p.conn()
 	return err
 }
 
@@ -68,11 +73,17 @@ func (p *Pool) Ping() error {
 // It is rare to Close a Pool, as the Pool handle is meant to be
 // long-lived and shared between many goroutines.
 func (p *Pool) Close() error {
-	p.mu.RLock()
-	defer p.mu.RUnlock()
-	if p.closed {
+	if atomic.LoadInt32(&p.closed) == poolIsClosed {
 		return nil
 	}
+
+	p.mu.Lock()
+	defer p.mu.Unlock()
+
+	if p.closed == poolIsClosed {
+		return nil
+	}
+	p.closed = poolIsClosed
 
 	for _, c := range p.conns {
 		err := c.Close()
@@ -84,14 +95,10 @@ func (p *Pool) Close() error {
 	return nil
 }
 
-func (p *Pool) conn() (*Connection, *pool.PoolConn, error) {
-	p.mu.RLock()
-
-	if p.closed {
-		p.mu.RUnlock()
-		return nil, nil, errPoolClosed
+func (p *Pool) conn() (*Connection, error) {
+	if atomic.LoadInt32(&p.closed) == poolIsClosed {
+		return nil, errPoolClosed
 	}
-	p.mu.RUnlock()
 
 	pos := atomic.AddInt32(&p.pointer, 1)
 	if pos == int32(len(p.conns)) {
@@ -99,7 +106,20 @@ func (p *Pool) conn() (*Connection, *pool.PoolConn, error) {
 	}
 	pos = pos % int32(len(p.conns))
 
-	return p.conns[pos], nil, nil
+	if p.conns[pos] == nil {
+		p.mu.Lock()
+		defer p.mu.Unlock()
+
+		if p.conns[pos] == nil {
+			var err error
+			p.conns[pos], err = NewConnection(p.host.String(), p.opts)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
+	return p.conns[pos], nil
 }
 
 // SetInitialPoolCap sets the initial capacity of the connection pool.
@@ -128,7 +148,7 @@ func (p *Pool) SetMaxOpenConns(n int) {
 
 // Exec executes a query without waiting for any response.
 func (p *Pool) Exec(ctx context.Context, q Query) error {
-	c, _, err := p.conn()
+	c, err := p.conn()
 	if err != nil {
 		return err
 	}
@@ -139,7 +159,7 @@ func (p *Pool) Exec(ctx context.Context, q Query) error {
 
 // Query executes a query and waits for the response
 func (p *Pool) Query(ctx context.Context, q Query) (*Cursor, error) {
-	c, _, err := p.conn()
+	c, err := p.conn()
 	if err != nil {
 		return nil, err
 	}
@@ -152,7 +172,7 @@ func (p *Pool) Query(ctx context.Context, q Query) (*Cursor, error) {
 func (p *Pool) Server() (ServerResponse, error) {
 	var response ServerResponse
 
-	c, _, err := p.conn()
+	c, err := p.conn()
 	if err != nil {
 		return response, err
 	}