+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## v6.1.0 - 2020-03-09
+
+- Reworked and tested new connection pools with multiple queries per connection
+- Socket Read- and WriteTimeout replaced with context timeout
+- Mock assert fix
+- Connection pool fixed initial size
+- Changes added offsets
+
 ## v6.0.0 - 2019-12-22
 
 - 2.4 RethinkDB support